is added tests for reducer user
