+++ src/components/app/app.js
@@ -1,5 +1,5 @@
-import React from 'react';
-import MainPage from '../main-page/main-page.jsx';
+import React from "react";
+import MainPage from "../main-page/main-page.jsx";
 
 const App = () =>{
   return (
