+++ src/components/card-place/card-pace.e2e.test.js
@@ -1,7 +1,7 @@
-import React from 'react';
-import {shallow, configure} from 'enzyme';
-import CardPlace from './card-place.jsx';
-import Adapter from 'enzyme-adapter-react-16';
+import React from "react";
+import {shallow, configure} from "enzyme";
+import CardPlace from "./card-place.jsx";
+import Adapter from "enzyme-adapter-react-16";
 
 configure({adapter: new Adapter()});
