+++ src/components/card-place/card-place.jsx
@@ -1,5 +1,5 @@
-import React, {PureComponent} from 'react';
-import PropTypes from 'prop-types';
+import React, {PureComponent} from "react";
+import PropTypes from "prop-types";
 
 class CardPlace extends PureComponent {
   constructor(props) {
