+++ src/components/catalog/catalog.test.js
@@ -1,6 +1,6 @@
-import React from 'react';
-import {shallow, configure} from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
+import React from "react";
+import {shallow, configure} from "enzyme";
+import Adapter from "enzyme-adapter-react-16";
 
 import {Catalog} from "./catalog.jsx";
