+++ src/components/city-link/city-link.e2e.test.js
@@ -1,8 +1,8 @@
-import React from 'react';
-import {shallow, configure} from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
+import React from "react";
+import {shallow, configure} from "enzyme";
+import Adapter from "enzyme-adapter-react-16";
 
-import {CityLink} from './city-link.jsx';
+import {CityLink} from "./city-link.jsx";
 
 configure({adapter: new Adapter()});
