+++ src/components/login-form/login-form.jsx
@@ -1,5 +1,5 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import React from "react";
+import PropTypes from "prop-types";
 import {connect} from "react-redux";
 
 import {Operation} from "../../reducer/user/user";
