+++ src/components/map/map.e2e.test.js
@@ -1,6 +1,6 @@
-import React from 'react';
-import {Map} from './map.jsx';
-import renderer from 'react-test-renderer';
+import React from "react";
+import {Map} from "./map.jsx";
+import renderer from "react-test-renderer";
 
 const coordinates = [[52.3809553943508, 4.939309666406198]];
