+++ src/hocs/with-login-form/with-login-form.js
@@ -1,4 +1,4 @@
-import React, {PureComponent} from 'react';
+import React, {PureComponent} from "react";
 
 export const withLoginForm = (Component) => {
   class WithLoginForm extends PureComponent {
