+++ src/reducer/reducer.js
@@ -1,8 +1,8 @@
-import {combineReducers} from 'redux';
+import {combineReducers} from "redux";
 
-import {reducer as data} from './data/data';
-import {reducer as checkAuthorization} from './user/user';
-import NameSpace from './name-spaces';
+import {reducer as data} from "./data/data";
+import {reducer as checkAuthorization} from "./user/user";
+import NameSpace from "./name-spaces";
 
 
 export default combineReducers({
