+++ src/reducer/user/selectors.js
@@ -1,4 +1,4 @@
-import NameSpace from '../name-spaces';
+import NameSpace from "../name-spaces";
 
 const NAME_SPACE = NameSpace.USER;
