+++ src/reducer/user/user.js
@@ -1,12 +1,5 @@
-import Constants from './../../constants';
-
-const toModelUserDate = (userData) => ({
-  id: userData.id,
-  email: userData.email,
-  name: userData.name,
-  avatar: userData[`avatar_url`],
-  isPro: userData[`is_pro`],
-});
+import Constants from "./../../constants";
+import toModelUserDate from "./adapter";
 
 const initialState = {
   isAuthorizationRequired: false,
